add option to register multiple loggers
It is now possible to register multiple loggers, with different names.

This commit adds also the relevant convenience macros, with postfix `N`.

Only the main, default name's registration is enforced by default. The verification of other logger registrations are only enabled in case `STRICT_REGISTRATION` macro is defined. Though it's not a very expensive check, still thinking if there is anything else that should be cut out...